* Use nonsaturated dipole-proton cross section also when calculating   scattering amplitude for coherent dipole-nucleus scattering in   IPsat model if mode is set to IPSAT_MODE_NONSAT_P * Update documentation

+++ src/dipxs_ipsat.h
@@ -54,7 +54,7 @@ class Dipxs_IPSat : public Dipxs
         void Intialize();
         REAL Sigmap(REAL rsqr, REAL xbjork);
         REAL B_p;   
-        REAL mode;    
+        int mode;    
         bool factorize;
         gsl_integration_workspace *ft_workspace;
         gsl_integration_workspace *ft_workspace_proton;
